+++ app/src/main/java/com/example/androidcourseshpp/ui/screens/main/addcontacts/AddContactsContract.kt
@@ -1,9 +1,9 @@
 package com.example.androidcourseshpp.ui.screens.main.addcontacts
 
-import com.example.androidcourseshpp.ui.screens.main.addcontacts.model.UserItem
 import com.example.androidcourseshpp.ui.ViewEffect
 import com.example.androidcourseshpp.ui.ViewEvent
 import com.example.androidcourseshpp.ui.ViewState
+import com.example.androidcourseshpp.ui.screens.main.addcontacts.model.UserItem
 
 class AddContactsContract {
     sealed interface Event : ViewEvent {
@@ -16,7 +16,7 @@ class AddContactsContract {
         data class OnSearchBarTextChanged(val input: String) : Event
         data class SearchModeSwitched(val isSearchModeEnabled: Boolean) : Event
         data class OnAddContactClicked(
-            val userId: Int,
+            val userId: Long,
             val interruptSuccessLoading: () -> Unit,
             val interruptFailureLoading: () -> Unit
         ) : Event
